requirements.txt
opencv-python
torch==1.7.1+cu110
torchvision==0.8.2+cu110
numpy
matplotlib
pandas
Pillow
pycocotools